requirements.txt
dj-rest-auth==2.2.4
Django==4.0.5
django-admin-interface==0.19.1
django-allauth==0.51.0
django-colorfield==0.7.1
django-cors-headers==3.13.0
django-filer==2.2.1
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-sendgrid-v5==1.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
easy-thumbnails==2.8.1
gunicorn==20.1.0
Pillow==9.1.1
python-dotenv==0.20.0
sendgrid==6.9.7
whitenoise==6.2.0